Correccion de datos numericos en el script de HabilidadInmunidad

+++ Assets/Scripts/Jugadores/Habilidades/Support/Healer/HabilidadInmunidad.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class HabilidadInmunidad : HabilidadBase
 {
     [Header("Ajustes de Inmunidad")]
-    [SerializeField] private float duracion = 3f; // tiempo que dura el buff, se puede modificar igual en el inspector
+    [SerializeField] private int duracion = 3; // tiempo que dura el buff, se puede modificar igual en el inspector
 
     public override void Ejecutar(GameObject usuario, JugadorController aliadoCazador)
     {
